updated testing file names
