+++ __tests__/test components/logo.test.js
@@ -1,10 +1,9 @@
-import Project from '../../client/components/Project.jsx';
+import Logo from '../../client/components/Logo.jsx';
 import React from 'react';
 import { fireEvent, render, screen, waitFor, } from '@testing-library/react';
 
-
 test("Logo should be rendered", () => {
-    render(<Project />)
-    const projectCard = screen.getByTestId('project-test');
-    expect(projectCard).toBeInTheDocument();
+  render(<Logo />)
+  const svgLogo = screen.getByTestId('logo-svg-test');
+  expect(svgLogo).toBeInTheDocument();
 })
