+++ __tests__/test components/project.test.js
@@ -1,9 +1,10 @@
-import Logo from '../../client/components/Logo.jsx';
+import Project from '../../client/components/Project.jsx';
 import React from 'react';
 import { fireEvent, render, screen, waitFor, } from '@testing-library/react';
 
+
 test("Logo should be rendered", () => {
-  render(<Logo />)
-  const svgLogo = screen.getByTestId('logo-svg-test');
-  expect(svgLogo).toBeInTheDocument();
+    render(<Project />)
+    const projectCard = screen.getByTestId('project-test');
+    expect(projectCard).toBeInTheDocument();
 })
